pyproject.toml
[project]
name = "lighting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dirigera>=1.1.9",
    "illumipy>=1.4.26",
    "matplotlib>=3.9.2",
    "numpy>=2.1.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
main = "lighting.lighting:main"

[tool]
uv.package = true
